任意の式の値計算を実装
